lldpd: Implement more setting options
With `configure system platform` you can configure the platform for the CDP protocol.
If you specify the flags for the legacy / non-LLDP protocols several times in succession, the protocol is also activated if no neighbor is found that speaks this protocol. With CDP this has the additional function to indicate the protocol version.
Furthermore, the option to use two flags to send less information has been added (`-k`, `-i`).
Further details can be found in the man pages.

Signed-off-by: Marek Küthe <m.k@mk16.de>